+++ src/main/java/com/arxyt/dominionsword/pomkotscompat/control/MechControlBridge.java
@@ -5,5 +5,4 @@ public interface MechControlBridge {
     MechControlFrame dominion$getControlFrame();
     void dominion$queueDriverInput(short bits);
     short dominion$getLastAppliedDriverInput();
-    short dominion$getQueuedDriverInput();
 }